fix(signing): rename mobile -> MobileSigning, Signing -> ReactNativeSigning
