+++ packages/signing/android/src/main/java/org/iota/MobileSigning/Interface.java
@@ -1,4 +1,4 @@
-package org.iota.mobile;
+package org.iota.MobileSigning;
 
 public class Interface {
     static { System.loadLibrary("dummy"); }
